[client] fix handling of custom scalar inputs (#1266)
Update client generation logic to correctly annotate custom scalars when they are used as input fields. Due to the extra wrapper layer, optional input functionality is not supported with custom scalars at this point.

This PR also fixes client generation logic to map GraphQL Floats to Kotlin Double (instead of a Float).

Related:

* partially resolves https://github.com/ExpediaGroup/graphql-kotlin/issues/1263

+++ plugins/client/graphql-kotlin-client-generator/src/test/data/generator/input_self_reference/inputs/ComplexArgumentInput.kt
@@ -1,7 +1,7 @@
 package com.expediagroup.graphql.generated.inputs
 
 import com.expediagroup.graphql.client.Generated
-import kotlin.Float
+import kotlin.Double
 
 /**
  * Self referencing input object
@@ -11,11 +11,11 @@ public data class ComplexArgumentInput(
   /**
    * Maximum value for test criteria
    */
-  public val max: Float? = null,
+  public val max: Double? = null,
   /**
    * Minimum value for test criteria
    */
-  public val min: Float? = null,
+  public val min: Double? = null,
   /**
    * Next criteria
    */
