+++ plugins/client/graphql-kotlin-client-generator/src/test/data/jackson/variables/inputs/SimpleArgumentInput.kt
@@ -3,7 +3,7 @@ package com.expediagroup.graphql.generated.inputs
 import com.expediagroup.graphql.client.Generated
 import com.expediagroup.graphql.client.jackson.types.OptionalInput
 import com.expediagroup.graphql.client.jackson.types.OptionalInput.Undefined
-import kotlin.Float
+import kotlin.Double
 import kotlin.String
 
 /**
@@ -14,11 +14,11 @@ public data class SimpleArgumentInput(
   /**
    * Maximum value for test criteria
    */
-  public val max: OptionalInput<Float> = OptionalInput.Undefined,
+  public val max: OptionalInput<Double> = OptionalInput.Undefined,
   /**
    * Minimum value for test criteria
    */
-  public val min: OptionalInput<Float> = OptionalInput.Undefined,
+  public val min: OptionalInput<Double> = OptionalInput.Undefined,
   /**
    * New value to be set
    */
